[docs-infra] Throw on non-constant exports; inline the test parser
Replace the console.warn with a thrown error when a metadata file mixes
non-constant exports with constants — the group replaces the file's
exports wholesale, so a stray export (including a constant widened off
its literal type) would be documented nowhere. Failing the build is
better than dropping it silently.

Inline parseSources into transformConstantGroup.test.ts and drop
parseSources.testUtils.ts, so the PR doesn't introduce a .testUtils.ts
file that the build ships as dead code.
